[refactor]: account feature move to data package
